+++ base/utils.c
@@ -950,8 +950,8 @@ static timerange* _get_matching_timerange(time_t test_time, timeperiod *tperiod)
 
 #ifdef TEST_TIMEPERIODS_A
 			printf("TYPE: %d\n", daterange_type);
-			printf("TEST:     %lu = %s", (unsigned long)test_time, ctime(&test_time));
-			printf("MIDNIGHT: %lu = %s", (unsigned long)midnight, ctime(&midnight));
+			printf("TEST:     %llu = %s", (unsigned long long)test_time, ctime(&test_time));
+			printf("MIDNIGHT: %llu = %s", (unsigned long long)midnight, ctime(&midnight));
 #endif
 
 			/* get the start time */
@@ -1027,8 +1027,8 @@ static timerange* _get_matching_timerange(time_t test_time, timeperiod *tperiod)
 				}
 
 #ifdef TEST_TIMEPERIODS_A
-			printf("START:    %lu = %s", (unsigned long)start_time, ctime(&start_time));
-			printf("END:      %lu = %s", (unsigned long)end_time, ctime(&end_time));
+			printf("START:    %llu = %s", (unsigned long long)start_time, ctime(&start_time));
+			printf("END:      %llu = %s", (unsigned long long)end_time, ctime(&end_time));
 #endif
 
 			/* start date was bad, so skip this date range */
@@ -1108,8 +1108,8 @@ static timerange* _get_matching_timerange(time_t test_time, timeperiod *tperiod)
 				}
 
 #ifdef TEST_TIMEPERIODS_A
-			printf("NEW START:    %lu = %s", (unsigned long)start_time, ctime(&start_time));
-			printf("NEW END:      %lu = %s", (unsigned long)end_time, ctime(&end_time));
+			printf("NEW START:    %llu = %s", (unsigned long long)start_time, ctime(&start_time));
+			printf("NEW END:      %llu = %s", (unsigned long long)end_time, ctime(&end_time));
 			printf("%d DAYS PASSED\n", days);
 			printf("DLST SHIFT:   %i\n", shift);
 #endif
@@ -1225,7 +1225,7 @@ static void _get_next_invalid_time(time_t pref_time, time_t *invalid_time, timep
 			if(potential_time > earliest_time) {
 				earliest_time = potential_time;
 #ifdef TEST_TIMEPERIODS_B
-				printf("    EARLIEST INVALID TIME: %lu = %s", (unsigned long)earliest_time, ctime(&earliest_time));
+				printf("    EARLIEST INVALID TIME: %llu = %s", (unsigned long long)earliest_time, ctime(&earliest_time));
 #endif
 				}
 			}
@@ -1237,7 +1237,7 @@ static void _get_next_invalid_time(time_t pref_time, time_t *invalid_time, timep
 			}
 		}
 #ifdef TEST_TIMEPERIODS_B
-		printf("    FINAL EARLIEST INVALID TIME: %lu = %s", (unsigned long)earliest_time, ctime(&earliest_time));
+		printf("    FINAL EARLIEST INVALID TIME: %llu = %s", (unsigned long long)earliest_time, ctime(&earliest_time));
 #endif
 
 	if (depth == max_depth)
@@ -1311,7 +1311,7 @@ void _get_next_valid_time(time_t pref_time, time_t *valid_time, timeperiod *tper
 			if(have_earliest_time == FALSE || potential_time < earliest_time) {
 				earliest_time = potential_time;
 #ifdef TEST_TIMEPERIODS_B
-				printf("    EARLIEST TIME: %lu = %s", (unsigned long)earliest_time, ctime(&earliest_time));
+				printf("    EARLIEST TIME: %llu = %s", (unsigned long long)earliest_time, ctime(&earliest_time));
 #endif
 				}
 			have_earliest_time = TRUE;
@@ -1323,7 +1323,7 @@ void _get_next_valid_time(time_t pref_time, time_t *valid_time, timeperiod *tper
 			for(temp_timeperiodexclusion = tperiod->exclusions; temp_timeperiodexclusion != NULL; temp_timeperiodexclusion = temp_timeperiodexclusion->next) {
 				_get_next_invalid_time(earliest_time, &earliest_time, temp_timeperiodexclusion->timeperiod_ptr);
 #ifdef TEST_TIMEPERIODS_B
-				printf("    FINAL EARLIEST TIME: %lu = %s", (unsigned long)earliest_time, ctime(&earliest_time));
+				printf("    FINAL EARLIEST TIME: %llu = %s", (unsigned long long)earliest_time, ctime(&earliest_time));
 #endif
 				}
 			}
@@ -2786,7 +2786,7 @@ int update_check_stats(int check_type, time_t check_time) {
 			check_statistics[check_type].bucket[x] = 0;
 		check_statistics[check_type].overflow_bucket = 0;
 #ifdef DEBUG_CHECK_STATS
-		printf("CLEARING ALL: TYPE[%d], CURRENT=%lu, LASTUPDATE=%lu\n", check_type, (unsigned long)current_time, (unsigned long)check_statistics[check_type].last_update);
+		printf("CLEARING ALL: TYPE[%d], CURRENT=%llu, LASTUPDATE=%llu\n", check_type, (unsigned long long)current_time, (unsigned long long)check_statistics[check_type].last_update);
 #endif
 		}
 
@@ -2868,7 +2868,7 @@ int generate_check_stats(void) {
 				check_statistics[check_type].bucket[x] = 0;
 			check_statistics[check_type].overflow_bucket = 0;
 #ifdef DEBUG_CHECK_STATS
-			printf("GEN CLEARING ALL: TYPE[%d], CURRENT=%lu, LASTUPDATE=%lu\n", check_type, (unsigned long)current_time, (unsigned long)check_statistics[check_type].last_update);
+			printf("GEN CLEARING ALL: TYPE[%d], CURRENT=%llu, LASTUPDATE=%llu\n", check_type, (unsigned long long)current_time, (unsigned long long)check_statistics[check_type].last_update);
 #endif
 			}
 
@@ -2884,7 +2884,7 @@ int generate_check_stats(void) {
 					break;
 
 #ifdef DEBUG_CHECK_STATS
-				printf("GEN CLEARING BUCKET %d, (NEW=%d, OLD=%d), CURRENT=%lu, LASTUPDATE=%lu\n", this_bucket, new_current_bucket, check_statistics[check_type].current_bucket, (unsigned long)current_time, (unsigned long)check_statistics[check_type].last_update);
+				printf("GEN CLEARING BUCKET %d, (NEW=%d, OLD=%d), CURRENT=%llu, LASTUPDATE=%llu\n", this_bucket, new_current_bucket, check_statistics[check_type].current_bucket, (unsigned long long)current_time, (unsigned long long)check_statistics[check_type].last_update);
 #endif
 
 				/* clear old bucket value */
@@ -2898,7 +2898,7 @@ int generate_check_stats(void) {
 			}
 #ifdef DEBUG_CHECK_STATS
 		else
-			printf("GEN NO CLEARING NEEDED: TYPE[%d], CURRENT=%lu, LASTUPDATE=%lu\n", check_type, (unsigned long)current_time, (unsigned long)check_statistics[check_type].last_update);
+			printf("GEN NO CLEARING NEEDED: TYPE[%d], CURRENT=%llu, LASTUPDATE=%llu\n", check_type, (unsigned long long)current_time, (unsigned long long)check_statistics[check_type].last_update);
 #endif
 
 		/* update last check time */
